requirements.txt
streamlit>=1.28.0
llama-index>=0.9.0
llama-index-llms-llama-cpp>=0.1.0
llama-index-embeddings-huggingface>=0.1.0
llama-index-vector-stores-faiss>=0.1.0
faiss-cpu>=1.7.4
transformers>=4.35.0
sentence-transformers>=2.2.2
pandas>=1.5.0
python-docx>=0.8.11
markdown>=3.5.0
psutil>=5.9.0